+++ tests/testthat/test_GPfit_regr_gpfit.R
@@ -2,8 +2,9 @@ skip_if_not_installed("GPfit")
 skip_on_os("windows")
 
 test_that("autotest", {
+  withr::local_seed(42)
   learner = lrn("regr.gpfit")
-expect_learner(learner)
+  expect_learner(learner)
   capture.output({
     result = run_autotest(learner, check_replicable = FALSE)
   })